fixed project names, added DanBoxContext, added interfaces and renamed controller
