+++ BLL/Facades/QuestionFacade.cs
@@ -1,0 +1,12 @@
+﻿using BLL.Interfaces;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace BLL.Facades
+{
+    class QuestionFacade: IQuestionFacade
+    {
+
+    }
+}
