+++ BLL/Interfaces/IQuestionRepository.cs
@@ -1,9 +1,9 @@
-﻿using BusinessLogicLayer.Models;
+﻿using BLL.Models;
 using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace BusinessLogicLayer.Interfaces
+namespace BLL.Interfaces
 {
     public interface IQuestionRepository
     {
