+++ BLL/Providers/QuestionProvider.cs
@@ -1,0 +1,12 @@
+﻿using BLL.Interfaces;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace BLL.Providers
+{
+    class QuestionProvider: IQuestionProvider
+    {
+
+    }
+}
