--- Tests/UnitTest1.cs
@@ -1,14 +1,0 @@
-using System;
-using Xunit;
-
-namespace Tests
-{
-    public class UnitTest1
-    {
-        [Fact]
-        public void Test1()
-        {
-
-        }
-    }
-}
